This contains all SHA's starting with
f0998ec3bc2b838a147bdc69f98097a4d290fa08
up to
131fdfd1a89a257094148c07c630f136752c703b

Approval for this merge patch constitutes approval and acceptance
of the Qt CLA for the above mentioned patch series.

+++ libpyside/pysideconversions.h
@@ -239,7 +239,8 @@ struct QFlagsConverter
 
     static inline T toCpp(PyObject* pyObj)
     {
-        long val = 0;
+        /* this was long. Needed int in Qt5 */
+        int val = 0;
         if (Shiboken::Enum::check(pyObj)) {
             val = Shiboken::Enum::getValue(pyObj);
         } else if (PyObject_TypeCheck(pyObj, Shiboken::SbkType<T>())) {
